requirements.txt
django
djangorestframework
django-cors-headers
SQLAlchemy
celery
django-celery-beat
django_celery_results
requests
filehash
tqdm